Refactor to modularize monitor
